pyproject.toml
[tool.poetry]
name = "strava-gear"
version = "0.1.0"
description = ""
authors = ["Blake Peterman <blake@peterman.io>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
Django = "^4.0.2"
psycopg2 = "^2.9.3"
django-extensions = "^3.1.5"
djangorestframework = "^3.13.1"
django-cors-headers = "^3.11.0"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
isort = "^5.10.1"
flake8 = "^4.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
